requirements.txt
aiohttp~=3.13.1
kurigram>=2.2.13
pillow~=12.0.0
psutil~=7.1.2
pymongo>=4.15.3
pytgcrypto>=1.2.11
py-tgcalls~=2.2.8
py-yt-search~=0.3
python-dotenv~=1.1.1
yt-dlp>=2025.10.22
